+++ src/channel.ts
@@ -29,7 +29,6 @@ import {
   MessageFilters,
   MessageResponse,
   MuteChannelAPIResponse,
-  PaginationOptions,
   PartialUpdateChannel,
   PartialUpdateChannelAPIResponse,
   QueryMembersOptions,
@@ -49,6 +48,7 @@ import {
   DefaultGenerics,
   PinnedMessagePaginationOptions,
   PinnedMessagesSort,
+  MessagePaginationOptions,
 } from './types';
 import { Role } from './permissions';
 
@@ -773,7 +773,7 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
    */
   async getReplies(
     parent_id: string,
-    options: PaginationOptions & { user?: UserResponse<StreamChatGenerics>; user_id?: string },
+    options: MessagePaginationOptions & { user?: UserResponse<StreamChatGenerics>; user_id?: string },
   ) {
     const data = await this.getClient().get<GetRepliesAPIResponse<StreamChatGenerics>>(
       this.getClient().baseURL + `/messages/${parent_id}/replies`,
@@ -862,6 +862,10 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
     if (message.user?.id === this.getClient().userID) return false;
     if (message.user?.id && this.getClient().userMuteStatus(message.user.id)) return false;
     if (message.type === 'system') return false;
+
+    // Return false if channel doesn't allow read events.
+    if (Array.isArray(this.data?.own_capabilities) && !this.data?.own_capabilities.includes('read-events'))
+      return false;
 
     if (this.muteStatus().muted) return false;